Add files via upload
added input display on this version 
and the time base is now uses adjustable threw the up and down arrow keys
and it now has a full range marker so you can see where the new time results is being displayed
the index numbers are padded so they do not jump around on the status line
the traditional speed clocking. the indexes work and the trace does not go off the screen when you change speeds
the bit input display is extremely nice! it displays a block of pixels that represent the real time input activity on gpio pin 2 yet it is delayed by the sample time since it would slow the input chronically to attem,pt to display it real time. the bit display shows the first 12 thousand bits of input "per time base period" detected on gpio pin 2 bits of input. if there is no input , the bit block is blank(representing all zeros the time base is adjustable threw the up and down arrow keys yet if you go below 1 second, it will most likely input many times the 12k bits that it can display,, and it only samples the keyboard one time period so it might take a long time to respond if you adjust the time base to below 10 seconds..(10 seconds on 10 seconds etc..) yet it starts out on .1 second i think,,, i think.. lol.. have fun