Add files via upload
uses multiple pin write to write successive on and off to the clock pin speed it up a little.. but i did discover writing to the same pin with multi pin write you can generate any serial pattern at 3mbps yet to only to one pin at a time . i was only able to legitimize writing on 2 instruction cycles (on and off)to the clock pin with multi-pin list. so a list of 2 pins. yet 3 million pin value changes per second are possible. so it prove that it is possible to send 3mbps yet only out 1 pin at a time